renamed dirs and files and updated imports
