+++ docs/spec_sheet_images/sprite_directions.png
