+++ game/managers/camera_manager.py
